# packages/contracts/openapi/analysis-service.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_ReportRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Всего записей, удовлетворяющих фильтру",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ReportRead]",
        "type": "object"
      },
...
      "ReportCreate": {
        "properties": {
          "object_id": {
            "format": "uuid",
            "title": "Object Id",
            "type": "string"
          },
          "period_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Первый местный день периода; по умолчанию — неделя до конца",
            "title": "Period From"
          },
          "period_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Последний местный день включительно; по умолчанию — день as_of",
            "title": "Period To"
          }
        },
        "required": [
          "object_id"
        ],
        "title": "ReportCreate",
        "type": "object"
      },
      "ReportCreated": {
        "properties": {
          "as_of": {
            "description": "Момент анализа, чьи выводы вошли в отчёт",
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence_images": {
            "description": "Снимков-доказательств вставлено",
            "title": "Evidence Images",
            "type": "integer"
          },
          "evidence_missing": {
            "description": "Снимков не вставлено; причины — в разделе «Ограничения»",
            "title": "Evidence Missing",
            "type": "integer"
          },
          "generated_on": {
            "description": "Местный день формирования — часть ключа",
            "format": "date",
            "title": "Generated On",
            "type": "string"
          },
          "key": {
            "description": "Ключ файла в бакете `reports`; им же читается ссылка",
            "title": "Key",
            "type": "string"
          },
          "object_id": {
            "format": "uuid",
            "title": "Object Id",
            "type": "string"
          },
          "period_from": {
            "format": "date",
            "title": "Period From",
            "type": "string"
          },
          "period_to": {
            "format": "date",
            "title": "Period To",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "summary_generated_by": {
            "description": "`TEMPLATE` или `LLM` — источник резюме",
            "title": "Summary Generated By",
            "type": "string"
          },
          "url": {
            "description": "Presigned-ссылка на S3_PUBLIC_ENDPOINT; живёт S3_PRESIGN_TTL_S",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "key",
          "object_id",
          "period_from",
          "period_to",
          "generated_on",
          "created_at",
          "size_bytes",
          "url",
          "as_of",
          "summary_generated_by",
          "evidence_images",
          "evidence_missing"
        ],
        "title": "ReportCreated",
        "type": "object"
      },
      "ReportRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "generated_on": {
            "description": "Местный день формирования — часть ключа",
            "format": "date",
            "title": "Generated On",
            "type": "string"
          },
          "key": {
            "description": "Ключ файла в бакете `reports`; им же читается ссылка",
            "title": "Key",
            "type": "string"
          },
          "object_id": {
            "format": "uuid",
            "title": "Object Id",
            "type": "string"
          },
          "period_from": {
            "format": "date",
            "title": "Period From",
            "type": "string"
          },
          "period_to": {
            "format": "date",
            "title": "Period To",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "url": {
            "description": "Presigned-ссылка на S3_PUBLIC_ENDPOINT; живёт S3_PRESIGN_TTL_S",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "key",
          "object_id",
          "period_from",
          "period_to",
          "generated_on",
          "created_at",
          "size_bytes",
          "url"
        ],
        "title": "ReportRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Сервис сверки план-факт: отклонения D1–D10 с объяснением, прогресс, SPI и прогноз. Каждое число сопровождается фактами, на которых оно построено.",
    "title": "analysis-service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analysis/reports": {
      "get": {
        "operationId": "list_reports_api_v1_analysis_reports_get",
        "parameters": [
          {
            "in": "query",
            "name": "object_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ReportRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Отчёты объекта, новые сверху",
        "tags": [
          "Отчёты"
        ]
      },
      "post": {
        "description": "Период — местные дни календаря объекта включительно. Формирование синхронное: ответ приходит с готовой ссылкой. Отчёт того же периода, сформированный в тот же день, заменяется.",
        "operationId": "create_report_api_v1_analysis_reports_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Сформировать PDF-отчёт за период",
        "tags": [
          "Отчёты"
        ]
      }
    },
    "/api/v1/analysis/reports/{key}": {
      "get": {
        "description": "`key` — ключ из списка, вида `{object_id}/{дата}-{начало}_{конец}.pdf`.",
        "operationId": "get_report_api_v1_analysis_reports__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ссылка на готовый отчёт",
        "tags": [
          "Отчёты"
        ]
      }
    },
...
  }
}